Incluindo consumo de ClientRegistred

+++ src/Bank.Core/Messages/Integration/ClientRegistredEvent.cs
@@ -3,10 +3,10 @@ namespace Bank.Core.Messages.Integration;
 public class ClientRegistredEvent : Event
 {
     public Guid ClientId { get; private set; }
-    public float SolicitedLimit { get; private set; }
+    public decimal SolicitedLimit { get; private set; }
     public string Document { get; private set; }
     
-    public ClientRegistredEvent(Guid clientId, float solicitedLimit, string document)
+    public ClientRegistredEvent(Guid clientId, decimal solicitedLimit, string document)
     {
         AggregateId = clientId;
         ClientId = clientId;
